Added Middeware 'admin', 'writer', settled 'avatar' and some other features

+++ app/Http/Controllers/TagController.php
@@ -99,8 +99,11 @@ class TagController extends Controller
      */
     public function destroy(Tag $tag)
     {
+        if ($tag->posts->count() > 0) {
+            session()->flash('error', 'Tag have related posts. Can\'t be deleted!');
+            return redirect()->back();
+        }
         $tag->delete();
-
         session()->flash('warning', 'tag deleted successfully!');
         return redirect(route('tags.index'));
     }
